Grammar: Fix scope for functions ending with apostrophe
Before this change a function that had the same name as a standard
function except that it ended with a `'` e.g. `flip'`, had a partially
wrong scope. The part before the apostrophe was reported as
`support.function.prelude.haskell`, while the apostrophe had the scope
`source.haskell`. Something like `flip'` is a valid function name in
Haskell. This means the scope of the whole function name should be the
same. After this commit function definitions like `flip'` or `sum'`
have the scope `source.haskell`, just like any other custom function
definition.