Remove usage of character class intersections
This is due to a limitation of the Github implementation of the grammar parser. Github uses PCRE instead of the Onigmo engine TextMate uses. The syntax features of each are very similar but PCRE does not implement intersections in character classes (`[a-z&&[^aeiou]]`).

I've removed all uses of intersections in favor of more conventional look-arounds.